+++ apiphany-httpclient5/src/main/java/org/apiphany/client/http/ApacheHC5HttpExchangeClient.java
@@ -1,11 +1,13 @@
 package org.apiphany.client.http;
 
 import java.io.File;
+import java.io.InputStream;
 import java.io.Serializable;
 import java.net.URI;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Supplier;
 
 import org.apache.hc.client5.http.classic.methods.HttpDelete;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
@@ -32,7 +34,6 @@ import org.apiphany.ApiResponse;
 import org.apiphany.client.ClientProperties;
 import org.apiphany.client.ExchangeClient;
 import org.apiphany.header.Headers;
-import org.apiphany.header.MapHeaderValues;
 import org.apiphany.http.ApacheHC5Entities;
 import org.apiphany.http.ContentEncoding;
 import org.apiphany.http.HttpContentType;
@@ -40,6 +41,8 @@ import org.apiphany.http.HttpException;
 import org.apiphany.http.HttpHeader;
 import org.apiphany.http.HttpMethod;
 import org.apiphany.http.HttpStatus;
+import org.apiphany.io.InputStreamSupplier;
+import org.apiphany.json.JsonBuilder;
 import org.apiphany.lang.Strings;
 import org.apiphany.lang.collections.Lists;
 import org.apiphany.lang.collections.Maps;
@@ -167,7 +170,6 @@ public class ApacheHC5HttpExchangeClient extends AbstractHttpExchangeClient {
 			HttpEntity httpEntity = createHttpEntity(apiRequest);
 			httpUriRequest.setEntity(httpEntity);
 		}
-
 		return httpUriRequest;
 	}
 
@@ -181,13 +183,32 @@ public class ApacheHC5HttpExchangeClient extends AbstractHttpExchangeClient {
 	 */
 	protected <T> HttpEntity createHttpEntity(final ApiRequest<T> apiRequest) {
 		T body = apiRequest.getBody();
-		String contentTypeValue = Lists.first(MapHeaderValues.get(HttpHeader.CONTENT_TYPE, apiRequest.getHeaders()));
+		List<String> contentTypeValues = getHeaderValues(HttpHeader.CONTENT_TYPE, apiRequest.getHeaders());
+		String contentTypeValue = Lists.first(contentTypeValues);
 		ContentType contentType = Nullables.apply(contentTypeValue, ct -> ContentType.parse(ct).withCharset(apiRequest.getCharset()));
+		return createHttpEntity(apiRequest, body, contentType);
+	}
+
+	/**
+	 * Creates an appropriate {@link HttpEntity} based on the request body type and content type.
+	 *
+	 * @param <T> request body type
+	 *
+	 * @param apiRequest API request object
+	 * @param body request body
+	 * @param contentType content type of the request body
+	 * @return HTTP entity object
+	 */
+	private static <T> HttpEntity createHttpEntity(final ApiRequest<T> apiRequest, final T body, final ContentType contentType) {
 		return switch (body) {
 			case String str -> HttpEntities.create(str, contentType);
 			case byte[] bytes -> HttpEntities.create(bytes, contentType);
+			case InputStream is -> ApacheHC5Entities.create(is, contentType);
+			case InputStreamSupplier iss -> ApacheHC5Entities.create(iss.get(), contentType);
+			case Supplier<?> supplier -> createHttpEntity(apiRequest, JavaObjects.cast(supplier.get()), contentType);
 			case File file -> HttpEntities.create(file, contentType);
 			case Serializable serializable -> HttpEntities.create(serializable, contentType);
+			case Object obj when isJson(apiRequest) -> HttpEntities.create(JsonBuilder.toJson(body), contentType);
 			default -> HttpEntities.create(Strings.safeToString(body), contentType);
 		};
 	}
@@ -206,13 +227,14 @@ public class ApacheHC5HttpExchangeClient extends AbstractHttpExchangeClient {
 	protected <T, U> ApiResponse<U> buildResponse(final ApiRequest<T> apiRequest, final ClassicHttpResponse response) {
 		HttpEntity httpEntity = response.getEntity();
 		HttpStatus httpStatus = HttpStatus.fromCode(response.getCode());
-
 		Map<String, List<String>> headers = Nullables.whenNotNull(response.getHeaders(), ApacheHC5HttpExchangeClient::toHttpHeadersMap);
 
-		List<String> encodings = getHeaderValuesChain().get(HttpHeader.CONTENT_ENCODING, headers);
+		List<String> encodings = getHeaderValues(HttpHeader.CONTENT_ENCODING, headers);
 		U responseBody = ContentEncoding.decodeBody(getResponseBody(apiRequest, httpEntity), ContentEncoding.parseAll(encodings));
 
-		HttpContentType contentType = HttpContentType.from(httpEntity.getContentType(), httpEntity.getContentEncoding());
+		List<String> contentTypes = getHeaderValues(HttpHeader.CONTENT_TYPE, headers);
+		HttpContentType contentType = HttpContentType.parse(contentTypes);
+
 		if (httpStatus.isError()) {
 			throw new HttpException(httpStatus, StringHttpContentConverter.from(responseBody, contentType));
 		}